% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fls.fit}
\alias{fls.fit}
\title{FLS fit}
\usage{
fls.fit(X, y, mu, smooth = FALSE)
}
\arguments{
\item{X}{design matrix of dimensuin \eqn{n * K}.}

\item{y}{vector of observations of length \eqn{n}.}

\item{mu}{parameter controling relative weight of sum of dynamic errors (\eqn{r_D^2}) vs sums of squared residual measurement errors (\eqn{r_M^2}).}

\item{smooth}{logical. If TRUE, a smoothed estimate is provided.}
}
\value{
A \eqn{n * K} matrix coefficient estimates.
}
\description{
This function is used to fit Time-Varying Linear Regression via Flexible least squares (FLS) as discribed in R. Kalaba and L. Tesfatsion (1989).
}
\references{
\insertRef{KALABA19891215}{fls}
}
