% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fls.R
\name{fls}
\alias{fls}
\title{Add together two numbers}
\usage{
fls(X, y, mu = 1, smooth = TRUE)
}
\arguments{
\item{y}{A number}

\item{x}{A number}
}
\value{
The sum of \code{x} and \code{y}
}
\description{
Add together two numbers
}
\examples{
add(1, 1)
add(10, 1)

}
